final changes

+++ BangazonAPI/Controllers/DepartmentsController.cs
@@ -49,7 +49,7 @@ namespace BangazonAPI.Controllers
                         {
                             Id = reader.GetInt32(reader.GetOrdinal("Id")),
                             Name = reader.GetString(reader.GetOrdinal("Name")),
-                            Budget = reader.GetDecimal(reader.GetOrdinal("Budget")),
+                            Budget = reader.GetInt32(reader.GetOrdinal("Budget")),
                             // You might have more columns
                         };
 
@@ -87,7 +87,7 @@ namespace BangazonAPI.Controllers
                         {
                             Id = reader.GetInt32(reader.GetOrdinal("Id")),
                             Name = reader.GetString(reader.GetOrdinal("Name")),
-                            Budget = reader.GetDecimal(reader.GetOrdinal("Budget")),
+                            Budget = reader.GetInt32(reader.GetOrdinal("Budget")),
                             // You might have more columns
                         };
                     }
